move setting nodes to subdir
